feat(#707) - fix attempting to delete not existing models

+++ application/app/Sync/Repositories/InstitutionUserRepository.php
@@ -80,8 +80,9 @@ class InstitutionUserRepository implements CachedEntityRepositoryInterface
 
     public function delete(string $id): void
     {
-        $obj = $this->getBaseModel()->find($id);
-        $obj->delete();
+        if ($obj = $this->getBaseModel()->find($id)) {
+            $obj->delete();
+        }
     }
 
     public function deleteNotSynced(): void
